Prepare the namespace for the v4 API
