+++ Gemfile
@@ -1,6 +1,6 @@
 source 'https://rubygems.org'
 
-# Specify your gem's dependencies in convertkit-ruby.gemspec
+# Specify your gem's dependencies in convertkit_v4-ruby.gemspec
 gemspec
 
 gem 'dotenv', '~> 2.1', '>= 2.1.1'
