+++ lib/convertkit_v4/client/account.rb
@@ -1,4 +1,4 @@
-module Convertkit
+module ConvertkitV4
   class Client
     module Account
       def account
