+++ lib/convertkit_v4/client/custom_fields.rb
@@ -1,4 +1,4 @@
-module Convertkit
+module ConvertkitV4
   class Client
     module CustomFields
