+++ lib/convertkit_v4/client/sequences.rb
@@ -1,4 +1,4 @@
-module Convertkit
+module ConvertkitV4
   class Client
     module Sequences
       def sequences
