+++ lib/convertkit_v4/client/subscribers.rb
@@ -1,4 +1,4 @@
-module Convertkit
+module ConvertkitV4
   class Client
     module Subscribers
       def subscribers(options = {})
