+++ lib/convertkit_v4/client/tags.rb
@@ -1,4 +1,4 @@
-module Convertkit
+module ConvertkitV4
   class Client
     module Tags
       def tags
