+++ lib/convertkit_v4/client/webhooks.rb
@@ -1,4 +1,4 @@
-module Convertkit
+module ConvertkitV4
   class Client
     module Webhooks
       def webhooks
