+++ lib/convertkit_v4/version.rb
@@ -1,3 +1,3 @@
-module Convertkit
+module ConvertkitV4
   VERSION = "1.0.0"
 end
