+++ spec/convertkit_v4/client/tags_spec.rb
@@ -1,16 +1,16 @@
 require "spec_helper"
 require "securerandom"
 
-module Convertkit
+module ConvertkitV4
   class Client
     describe Tags do
       before do
-        Convertkit.configure do |config|
+        ConvertkitV4.configure do |config|
           config.api_secret = ENV["API_SECRET"]
           config.api_key = ENV["API_KEY"]
         end
 
-        @client = Convertkit::Client.new
+        @client = ConvertkitV4::Client.new
       end
 
       describe "#tags" do
